refactor: API 리퀘스트 로직 훅으로 분리
